Remove gateway feature flags from model and http
