+++ src/gateway/sharding/shard_queuer.rs
@@ -37,6 +37,7 @@ use crate::model::gateway::{GatewayIntents, ShardInfo};
 /// A shard queuer instance _should_ be run in its own thread, due to the blocking nature of the
 /// loop itself as well as a 5 second thread sleep between shard starts.
 pub struct ShardQueuer {
+    pub(super) token: SecretString,
     /// A copy of [`Client::data`] to be given to runners for contextual dispatching.
     ///
     /// [`Client::data`]: crate::Client::data
@@ -215,7 +216,7 @@ impl ShardQueuer {
         };
         let mut shard = Shard::new(
             Arc::clone(&self.ws_url),
-            self.http.token(),
+            self.token.clone(),
             shard_info,
             self.intents,
             self.presence.clone(),
