+++ src/model/event.rs
@@ -10,13 +10,11 @@ use nonmax::NonMaxU64;
 use serde::de::Error as DeError;
 use serde::Serialize;
 use strum::{EnumCount, IntoStaticStr, VariantNames};
-#[cfg(feature = "gateway")]
-use tracing::{debug, warn};
 
 use crate::constants::Opcode;
 use crate::internal::utils::lending_for_each;
 use crate::model::prelude::*;
-use crate::model::utils::{deserialize_val, remove_from_map};
+use crate::model::utils::remove_from_map;
 
 /// Requires no gateway intents.
 ///
@@ -1332,21 +1330,5 @@ impl Event {
     #[must_use]
     pub fn name(&self) -> &'static str {
         self.into()
-    }
-
-    #[cfg(feature = "gateway")]
-    pub(crate) fn deserialize_and_log(map: JsonMap, original_str: &str) -> Result<Self> {
-        deserialize_val(Value::Object(map)).map_err(|err| {
-            let err_dbg = format!("{err:?}");
-            if let Some((variant_name, _)) =
-                err_dbg.strip_prefix(r#"Error("unknown variant `"#).and_then(|s| s.split_once('`'))
-            {
-                debug!("Unknown event: {variant_name}");
-            } else {
-                warn!("Err deserializing text: {err_dbg}");
-            }
-            debug!("Failing text: {original_str}");
-            Error::Json(err)
-        })
     }
 }
